refactor for readability
